+++ client/src/App.css
@@ -8,7 +8,7 @@
   flex-grow: 1;
   display: flex;
   flex-direction: column;
-  padding: 30px;
+  padding: 45px;
   background-color: #f9fafb;
   box-sizing: border-box;
   overflow-y: auto;
